create battle header component
